Introduce e_gdbus_templates_init_main_thread()
This is required to properly distinguish between EGDbusTemplates'
sync calls from the main thread and from a dedicated thread, because
the check with g_main_context_is_owner (g_main_context_default ())
doesn't work when there is no main loop running. The call to
e_gdbus_templates_init_main_thread() is good to do, but it is not
mandatory.