CamelIMAPXCommand: Make imapx_command_run_sync() more responsive to cancellations.
Connect a GCancellable::cancelled signal handler that broadcasts the
GCond that imapx_command_run_sync() is waiting on so it can proceed
immediately.

If imapx_command_complete() is called sometime later by the thread
executing the command, the GCond will broadcast again but no one will
be listening for it.