CamelVeeFolder: Process folder changes more efficiently.
CamelVeeFolder submits a new background job every time it receives a
CamelFolder::changed signal.  So the background jobs tend to pile up,
and there's no guarantee the jobs will be dispatched in the order
they were submitted since they're dispatched from a thread pool.

Instead, run one background job at a time, and have the job process
FolderChangedData from a queue until the queue is empty.  If another
CamelFolder::changed signal is received while the background job is
still running, the changes are simply be appended to the queue.