CamelIMAPXJob: Split out "folder_name" data.
Define standalone CreateFolderData and DeleteFolderData structs and
"free" functions.

Use camel_imapx_job_set_data() to attach an allocated struct to a
CamelIMAPXJob such that it will be freed along with the job.

Use camel_imapx_job_get_data() to obtain a pointer to the allocated
struct as needed, always with a NULL check for safety.