Bug 669487 - Fix up online notification (part 1)
Make E-D-S listen to the "start-offline" GSettings key (not GConf),
since Evolution no longer sets the GConf key.

Unfortunately we can't rely on the schema from Evolution, so we install
our own minimal schema that refers to the same DConf path as the
Evolution one.  This allows us to share state with no dependency issues.

This is a temporary solution for 3.4 only, to be removed in 3.5.