ENameSelectorEntry: Remove USER_QUERY_FIELDS macro.
No schema was ever defined for this GConf key so it couldn't have been
that important.  The symbol name itself is not namespaced properly (no
"E" prefix).  And it's inappropriate to use Evolution-specific keys in
E-D-S to begin with.

This is an API break.  There will be more along these lines.