EDataFactory: Keep only weak references to backends.
Have EDataFactory keep only weak references to the EBackend instances it
creates.  Then, when the last client connection to a backend closes, the
backend finalizes automatically.  No explicit action required.

This breaks the API slightly by renaming e_data_factory_get_backend() to
e_data_factory_ref_backend() to reflect the fact that a strong reference
is now returned to the caller, and the caller must call g_object_unref()
when finished with it.  The impact of the API break should be contained
within E-D-S.

This uses the new GWeakRef API in GLib 2.32 to ensure thread-safety.

The libebackend shared object name has already been bumped for 3.5.1.