Add EAsyncClosure.
EAsyncClosure provides a simple way to run an asynchronous function
synchronously without blocking a running GMainLoop or using threads.

Merging this API from the account-mgmt branch as it has proven stable
and quite convenient.  I've seen this pattern in use throughout GIO,
but to my knowledge it has never been formalized like this.