imap[x]: Remove shell command option from user interfaces.
Do not expose the IMAP shell command option in user interfaces.  I know
of only one user of this feature, and it's far too esoteric and eats up
too much screen real estate for Evolution's account editor.

The option is still supported, but the shell command needs to be set
directly in the account's key file now.  Key files are easier to read
and edit than the old GConf XML blobs so I think this is a reasonable
requirement.

Same goes for "Number of cached connections to use" in IMAPX: most users
won't know what that means, so change it in the key file if you need to.