EServerSideSource: Add a "write-directory" property.
Defines the directory where data source changes are to be written,
which may not be the same directory as it was originally read from.

Defaults to the value returned by e_server_side_source_get_user_dir().

This allows an ECollectionBackend to override the default directory to
keep key files for backend-created data sources separate from key files
for user-created data sources.