ECollectionBackendClass: Add dup_resource_id() method.
Returns the resource ID for a given child source.  A resource ID is
usually a unique, server-assigned identifier which is stored in the
child source in some backend-specific extension.  The method allows
ECollectionBackend to extract the ID without knowing exactly how it
is stored.  The method always returns a newly-allocated string.