e_book_backend_file_open: Fix directory existence test.
In e_book_backend_file_open() we seem to be bailing out if the data
directory does not exist, but yet we don't create the directory until
further down.

Not sure if this is correct, but it seems to me we only want to check
for the data directory if 'only_if_exists == TRUE'.