google: Add a new custom vCard property containing Google system groups
This can then be picked up by libfolks (and other interested clients) to
allow them to detect which system groups a Google contact is in. Previously,
they had to infer this from the categories the contact was in, which are
localised. This was tricky and fragile.

Helps: https://bugzilla.gnome.org/show_bug.cgi?id=676383