CamelStoreClass: Rename compare_folder_name to equal_folder_name.
CamelStoreClass has the following method:

    GCompareFunc compare_folder_name;

Both the method name and call signature imply this is a strcmp()-style
function where a zero return value means the values are equal.  But in
fact this method is used as an equality test where a zero return value
means the values are NOT equal.

Rename the method and change its call signature like so:

    GEqualFunc equal_folder_name;

This is an API break but clearly a necessary one.