Replace e_source_registry_server_queue_auth_session().
I realized ECollectionBackends are going to need a proper asynchronous
authentication function so they know when authentication is achieved.

e_source_registry_server_queue_auth_session() is a "fire-and-forget"
function with no way for the caller to know if and when authentication
is achieved.

The replacement function, e_source_registry_server_authenticate_sync(),
is a variation of e_source_registry_authenticate_sync() for use in the
registry service itself.  It takes an EAuthenticationSession instead of
an ESource.

Since this API has not yet seen a stable release, I'm going to forego
deprecating the old function and just remove it.  This is an API break,
and the libebackend soname has been bumped accordingly.