Bug 679808 - e_source_selector_set_primary_selection() has side-effects
In ESourceSelector parlance, the "primary selection" is the highlighted
tree view item.  A tree view item is "selected" if the check box cell is
ticked.  The two selection states are supposed to be orthogonal.

e_source_selector_set_primary_selection() has apparently for some time
been highlighting the requested item and also activating its check box.
That last part is an unwanted and undocumented side-effect.

This commit removes the side-effect and explicitly states in the API
documentation that the function does NOT alter the item's check box.