cache-reaper: Clean up abandoned data directories.
Cache reaper now cleans up abandoned data directories for local address
books and calendars, but more conservatively than for cache directories.

Whereas a cache directory is marked as trash as soon as its data source
is removed and then allowed to linger for one week, a data directory is
only moved to trash when the registry service starts, and is allowed to
linger for four weeks.  This should give ample time to recover the data
if needed.