Export the EDBusAuthenticator interface from an isolated thread.
This is similar to the problem I had with GDBusObjectManagerClient.
When a GDBusInterfaceSkeleton is exported, it grabs the thread-default
main context and emits method invocation signals from only that context.

Problem is when e_authenticator_try_password_sync() is called on an
EAuthenticationMediator, something may have pushed a different main
context and so the method invocation signal emissions are inhibited
and we eventually time out thinking the client is unresponsive.

The workaround is to export the GDBusInterfaceSkeleton from an
isolated thread where its signals cannot be inhibited.  The thread
runs its own main loop until the EAuthenticationMediator object is
finalized, at which point the thread terminates and is joined with.

This is the same as my initial attempt in commit [1] which I reverted,
except with improved thread-safety.

[1] commit: e13cb4e0ba820694f908fe39255ff8f7a6239038