CamelIMAPXServer: Add a "stream" property.
Protect the stream with a mutex and add camel_imapx_server_ref_stream().

Additionally, the stream pointer is now in the private structure, and
is explicitly passed to various parsing functions to ensure a reference
is held on the stream for the duration of a parser thread iteration.

This alters the signature of CamelIMAPUntaggedRespHandler.

I am, however, NOT changing libcamel's soname for these changes since
only evolution-kolab is affected and a soname bump at this point in the
development cycle is extremely disruptive to the rest of GNOME.