EContact: Added boxed type E_CONTACT_TYPE_ATTR_LIST and e_contact_field_type()
This boxed type is used in place of pointer types for properties
installed on EContact which are returned lists of strings.

This does not effect the behavior of e_contact_set()/e_contact_get(),
only it allows e_contact_field_type() to return a distinct type
for those fields, allowing automated code to introspect those useful
contact field types.

The e_contact_field_type() convenience function is used to fetch the GType
for a given field when used with EContact (this allows more automated
routines to be written for EContactFields).