Remove e_data_book_register_gdbus_object().
The first thing we do after creating a new EDataBook is export its D-Bus
interface.  This was a separate and failable operation.  If the exporting
fails, the EDataBook is useless.  Perfect use case for GInitable.

Now we pass the GDBusConnection and object path directly to
e_data_book_new(), and if the exporting fails the function sets a GError
and returns NULL.

This also introduces a couple accessor functions:

    e_data_book_get_connection()
    e_data_book_get_object_path()