Remove e_data_cal_view_register_gdbus_object().
Similar to EDataCal, the first thing we do after creating a new
EDataCalView is export its D-Bus interface.  This was a separate and
failable operation.  If the export fails, the EDataCalView is useless.
Perfect use case for GInitable.

Now we pass the GDBusConnection and object path directly to
e_data_cal_view_new(), and if exporting fails the function sets a
GError and returns NULL.

This also introduces a couple accessor functions:

    e_data_cal_view_get_connection()
    e_data_cal_view_get_object_path()