EAuthenticationSession: Close prompt before storing password.
Close our prompt before storing the password in the keyring.
If the keyring is locked, it will need to prompt the user for
a keyring password, but it can't do that if our password prompt
is still open since both prompts are system-modal.  Not sure what
would happen next; probably the store operation would either fail
or deadlock.