CamelIMAPXServer: Improve error handling around stopping IDLE.
imapx_stop_idle() returned a boolean but TRUE had an ambiguous
interpretation.  It either meant the DONE command was sent or an I/O
error occurred.  Define a new enum type (CamelIMAPXIdleStopResult) to
resolve the ambiguity.

Also imapx_command_idle_stop() returned a boolean but FALSE had an
ambiguous interpretation.  It either meant an I/O error occurred, or
there was no available CamelIMAPXStream from which to issue the DONE
command.  Resolve the ambiguity by forcing callers to explicitly pass
in a CamelIMAPXStream (and a GCancellable, while we're at it).