Adding test case proving there is a write-write race condition to fix
The test case creates a number of threads all trying to modify the
same contact in the same book each concurrently with a different field,
if the modification fails it retries. If all the fields (one for each
thread) are not properly set at the end of all thread execution, then
the test is a failure.