Revert "Revert "Bug #678400 - ECollectionBackend populates disabled accounts""
This reverts commit 99f0208d4005f2650729c9b63a629c17e1cd964b.

This reverts my revert, since I mistakenly thought it was breaking the
ESourceCollection API.  The solution is not ideal, but is fine as a
stop-gap measure for Evolution-Data-Server 3.6.x.

In 3.7.x I've added an e_source_registry_check_enabled() function and
swapped out most of the e_source_get_enabled() calls throughout Evo and
EDS, which obviates the need to keep enabled states for child sources
sychronized with parent sources at all times.