Bug 685986 - ESourceRegistry: Wait for signals after creating sources
e_source_registry_create_sources_sync() now waits for "object-added"
signals from its GDBusObjectManagerClient before returning.  This allows
the newly-created ESources to be obtained immediately after the function
returns.  Prior to this there was a small delay before the newly-created
ESources became available through the ESourceRegistry API.

There is a caveat to this.  If any of the scratch ESources reference a
parent UID which is not presently exported by the registry service, the
registry service will refuse to export the newly-created ESource until
its parent becomes available.

In this scenario, to avoid waiting forever for an "object-added" signal
that isn't coming, we place a limit on the wait time (two seconds).  If
this time limit expires, we return from the function regardless.