EBookBackendSqliteDB: Added e_book_backend_sqlitedb_get/set_revision()
In preperation for the migration, there needs to be a way to persist the
overall addressbook revision. This patch also increments the internal
SQLite DB format version and adds the new revision column to for main
folders table of an addressbook (and increments the stored DB version)
if the DB version is out of date.