Remove version number from ${privdatadir}.
In talking to some people at Red Hat about how to deploy pre-configured
Evolution accounts, one thing I suggested is to provide a package which
installs .source files to one of the system-wide locations:

    /usr/share/evolution-data-server-3.8/ro-sources
    /usr/share/evolution-data-server-3.8/rw-sources

But that's not going to work in general with the version number in the
path.  Upgrading E-D-S to, say, 3.10 would invalidate such a package.

Remove the version number from our ${privdatadir}.  It's not needed.