Add AllowAuthPromptAll() method to SourceManager interface.
Equivalent to calling AllowAuthPrompt() on each exported data source,
but cuts way down on the D-Bus traffic.  Since the registry service is
commonly activated during a desktop session startup, this can actually
clog up the session bus and slow down login.

This leaves the source-specific AllowAuthPrompt() method unused, but
there are still use cases for it which we haven't implemented yet, so
leave it be.