Avoid race conditions when concurrently writing contacts in the addressbook
This patch checks the incomming revision against the existing revision
for the local file backend and rejects the commit with a newly added
error code E_DATA_BOOK_STATUS_BAD_REVISION, which can be checked by the
client in order to retry the commit.

Note also, the revision format has changed to include the backend revision
counter which ensures revisions are unique (even when multiple revisions
occur inside the same second).