Added ESourceDirectAccess extension
The ESourceDirectAccess extension configures a backend for direct read
access mode by exposing the install path of the backend and it's backend
name.

This patch additionally moves the extensions from libebook to libebook-contacts
where it's visible from EDataBook.

EDataBook apis & EBookClient modified to use the new extension in order
to leverage the new extension instead of the hard coded variables.