EDataFactory: Add a "backend-created" signal.
Allows EDataFactory extensions to catch newly-created backends and
perform additional initialization on it.

The use case I have in mind is to move all the GOA initialization in
EDataBookFactory to the gnome-online-accounts module as an extension.

Note this uses one of the reserved pointer slots in EDataFactoryClass
and therefore does not break ABI.