Remove EBookBackendVCF.
Strolling through the commit history of this backend, I had to dig all
the way back to 2005 before I started seeing actual development work.
Since then the commits have all been code cleanups and adapting it to
backend API changes.

There is no way to create a VCF address book from Evolution, which means
this entire backend is dead code and we've been wasting our time with it.

Dump it.