Bug 688639 – SMTP: Correctly deal with 535 authentication error
Entering a wrong password, Evolution would fail to send messages using SMTP with the following error.

    Bad authentication response from server.

The log file of the mail server – Exim 4 in this case – contains the following error.

    cram_md5_server authenticator failed for myhost.example.org ([10.10.10.10]) [xx.xxx.xxx.xxx]: 535 Incorrect authentication data (set_id=joey)

Looking at the code after the rewrite in

    commit 68c269a66bd484768b9c1517e7dc2dd30f682485
    Author: Matthew Barnes <mbarnes@redhat.com>
    Date:   Sat Oct 15 09:51:35 2011 -0400

        smtp: Adapt to Camel's new authentication API.

present since git tag EVOLUTION_DATA_SERVER_3_3_1, accidentally changed the logic not acting correctly, when receiving a 535 authentication error response and therefore was overridden by the next code block and the variable `result` changed to `CAMEL_AUTHENTICATION_ERROR` instead of `CAMEL_AUTHENTICATION_REJECTED`, which resulted in the situation where the password was not reset and therefore the user not asked to reenter it.

Changing the logic, so the variable is not overridden successfully fixes the issue.