EBookBackendSqliteDB: Avoid errors on conflicting summaries
Added function introspect_summary() to derive the summary information
from a running database.

Now the initialization of the DB is a bit safer:
  o First collect configured summary
  o Create Database (if it doesnt exist) with the configured summary
  o Now derive the summary from the existing database

This ensures that inserts/selects run with the summary configuration
which matches the actual running database, regardless if the DB was
configured with the actual ESourceBackendSummarySetup passed or previously
configured with something else.