Remove introspection of libecal.
A great deal of the calendar API exposes data structures from libical.
libical is not a GObject library, and is therefore not introspectable.
This is unfortunate, but the API has been this way for over a decade.

The g-ir-scanner skips functions with data types it does not understand.
Therefore a great deal of the calendar API is omitted from introspection
and the resulting language bindings are not actually useful. This causes
confusion and disappointment, judging by the number of times I've had to
explain this to programmers in the #evolution channel.

Annotating the calendar API for introspection is still good practice.
But until this situation changes, let's not give the false impression
that E-D-S calendars are accessible from non-C languages by producing
a less-than-useful GIR file.

Remove the introspection.