Make ENamedParameters a fully boxed type
Introspection doesn't handle well typedefs as a general rule, but
typedefs for container types are particularly problematic.  This ends
up breaking the vala build, but the resulting .gir is also unusable by
bindings.

In this case, I think it's actually cleanest to make ENamedParameters
a fully opaque type even to C.  The parts of the Evolution code that
were peeking inside ENamedParameters really wanted a helper function
anyways to create a gchar **.

Therefore, this is a net code cleanup.

https://bugzilla.gnome.org/show_bug.cgi?id=689871