libecal: Move EDataCalCallStatus back to libedata-cal.
I'm encountering a build break in Evolution on an older environment.

It goes something like this...

   /* In X11/X.h */
   #define Success 0

   /* In libecal/e-cal-types.h */
   typedef enum { Success, ... } EDataCalCallStatus;

After pre-processing, the compiler gets

   typedef enum { 0, ... } EDataCalCallStatus;

which of course is a syntax error.

It's our own fault for violating our own namespace.  The enum definition
was moved from libedata-cal to libecal fairly recently (by me) to resolve
a circular dependency, but it appears libecal does not really require the
enum values.  Move the typedef back to libedata-cal as a quick Band-Aid.

The EDataCalCallStatus enum is redundant anyway and I plan to remove it.
Equivalent error values are provided by EClientError and ECalClientError,
and calendar backends should be using them directly.  There's no need for
separate but equal client and server-side error value definitions.

This is technically an API break, but I'm not going to bump the libecal
soname over this.  The enum definition is clearly intended for backends,
and because it was moved to libecal only recently, I'm fairly certain no
client applications are using it.