IMAPX: Support non-virtual Junk/Trash folders.
Some webmail clients designate specific IMAP folders as Junk and Trash,
such that when a message is deleted or determined to be spam, a copy of
the message is appended to the designated folder, the original message
is flagged as DELETED and the folder containing the original message is
immediately expunged.

This is significantly more expensive than simply flagging the message as
DELETED or JUNK.  But users that access their mail by both Evolution and
webmail frequently complain of seeing what appear to be duplicate Junk
and Trash folders in Evolution [1].

This commit allows IMAPX stores to be configured to mimic these webmail
clients and hide their Junk and Trash virtual folders.  IMAPX will mimic
the "copy-and-expunge" webmail client behavior each time it's commanded
to synchronize message flags with the server.

[1] So much so that it earned its own page in the user manual:
http://library.gnome.org/users/evolution/stable/mail-two-trash-folders.html.en