IMAPX: Fix DELETED flag sync when using real Trash folder.
After copying messages to a real Trash folder and then tagging them for
deletion, we immediately expunge.  Now the IMAPX expunge operation first
synchronizes message flags with the server, which is sensible.  But when
using a real Trash folder, this extra sync was causing deleted messages
to keep coming back from the grave like zombies.

The problem was when using a real Trash folder, the internal operation
that synchronizes flags first clears any DELETED flags.  Normally that's
what we want, except when we're about to expunge since expunging depends
on those DELETED flags.  So make an exception for this case.