IMAPX: Don't recheck flags before moving to real Junk/Trash.
Out of paranoia, I made imapx_folder_claim_move_to_real_trash_uids()
recheck that the queued message UIDs still had a DELETED flag before
moving them to a real Trash folder.  That turned out to be breaking
things because we _remove_ the DELETED flag when synchronizing, so
it's not going to still be set when we're ready to move messages.

Remove the uncertainty and rename public functions to reflect this:

  - camel_imapx_folder_maybe_move_to_real_junk()
  + camel_imapx_folder_add_move_to_real_junk()

  - camel_imapx_folder_maybe_move_to_real_trash()
  + camel_imapx_folder_add_move_to_real_trash()