IMAPX: Ask server to handle "body-contains" searches.
This isn't as sophisticated/convoluted as the version in the older IMAP
backend, but it still beats downloading the entire message as we've been
doing.  There's no caching of past search results.  I'm not convinced the
performance gain is worth the extra complexity.  If I'm wrong or I change
my mind, search result caching can always be added later.