ESourceRegistryServer: Add a "tweak-key-file" signal.
This signal is emitted from e_source_registry_server_load_file() just
prior to instantiating an EServerSideSource.  Signal handlers can tweak
the GKeyFile content as necessary and return TRUE to write the modified
content back to disk.

For the purposes of tweaking, it's easier to deal with a plain GKeyFile
than an ESource instance.  An ESource, for example, does not allow key
file groups to be removed.

The return value is cumulative.  If any signal handler returns TRUE,
the GKeyFile content is written back to disk.

Unfortunately this breaks the libebackend ABI because I forgot to add a
pool of reserved pointers to the ESourceRegistryServerClass.  Do so now,
since we have the opportunity.