IMAPX: Avoid duplicating messages in real Junk or Trash.
This is more a problem with Junk, but fixing both to keep the logic
consistent.  When synchronizing a real Junk folder, we were copying
all messages marked as "junk" from Junk to Junk -- duplicating them
each time.  After running for a couple days I caught my Junk folder
with thousands of duplicate messages.