google-backend: Populate [Authentication] group for contacts.
Set the Host to "www.google.com" and Method to either "OAuth2" or
"ClientLogin", depending on whether EOAuth2Support is available.

A temporary corner case occurs when using an older GNOME Online Accounts
version, which uses OAuth 1.0a instead of 2.0 for Google authentication.
In that case, the Method will say "ClientLogin" but the Google address
book backend will ignore it and use its EGDataGoaAuthorizer.