Merge ECalBackendFileStore into ECalBackendStore.
ECalBackendStore is an abstract base class [1], and ECalBackendFileStore
is a subclass that stores calendar components in an iCalendar file.  But
ECalBackendFileStore is the only subclass in existence, so the abstract
base class is unnecessary.

Have ECalBackendStore implement ECalBackendFileStore by default and drop
ECalBackendFileStore as a separate class.  The operations are written as
virtual class methods, so we can still override the class methods from a
new subclass if a need should arise.

New functions:

    e_cal_backend_store_new()

Removed functions:

    e_cal_backend_file_store_new()

[1] Or at least is supposed to be abstract.  Someone forgot to define
    the GType with G_DEFINE_ABSTRACT_TYPE.